100-prime_factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -6,13 +6,13 @@
 */
 int main(void)
 {
-long long int i, n = 612852475143;
+unsigned long int i, n = 612852475143;
 
 for (i = 3; i < 782849; i = i + 2)
 {
 while ((n % i == 0) && (n != i))
 n = n / i;
 }
-printf("%llu\n", n);
+printf("%lu\n", n);
 return (0);
 }
